pull new class out
